icrst username and password; changed annotation for tables hospital and region by specifying the table name with initial lower case. Using Long instead of int for hospital and region ids

+++ clubfootReg/java/edu/uiowa/icr/controllers/HospitalController.java
@@ -71,9 +71,9 @@ public class HospitalController {
 	@RequestMapping(value = "/edithospital", method = RequestMethod.GET) // In this instance we receive the jsp view "edit_evaluator" and get command to view the page
 	public ModelAndView editHospitalForm(HttpServletRequest request) {
 		System.out.println("THE ID OF THE HOSP TO BE EDITED: "+request.getParameter("Id")); 
-	    int hospitalId = Integer.parseInt(request.getParameter("id"));	// This parses an int from the request object's getParameter method for "id"
+	    Long hospitalId = Long.parseLong(request.getParameter("id"));	// This parses a Long from the request object's getParameter method for "id"
 	    Hospital h = hospitalRepo.getHospital(hospitalId);	// Instantiation of Hospital based on the integer id value of the hospital
-	    ModelAndView model = new ModelAndView("edithospital");	// Instantition of the Model and view built in method and passes edit_hospital view to it
+	    ModelAndView model = new ModelAndView("edithospital");	// Instantiation of the Model and view built in method and passes edit_hospital view to it
 	    model.addObject("hospital", h); // Adds an object to the model called hospital and passes the object h to it for the edit functionality
 	    return model; // Returns the model for the user to edit
 	}
@@ -89,7 +89,7 @@ public class HospitalController {
 	
 	
 	@ModelAttribute("regionList")
-	public Map<Integer, String> populateRegionSelect() {
+	public Map<Long, String> populateRegionSelect() {
 	    return hospitalRepo.getAllRegions();
 	}
