+++ clubfootReg/java/edu/uiowa/icr/implementation/inter/HospitalRepo.java
@@ -14,15 +14,15 @@ import edu.uiowa.icr.models.ReportsHospital;
  * Interface for the Hospital Repository to synchronize data to the database via the following interface methods.
  */
 public interface HospitalRepo {
-	Hospital getHospital1(int id);
+	Hospital getHospital1(Long id);
 	void addHospital(Hospital hospital); // addHospital method to add the model data to the database
 	void setDataSource(DataSource ds); // Sets up the dataSource for the database synch
-	Hospital getHospital(int id); // getHospital method that passes the int value for the id
-	void deleteHospital(int id); // DeleteHospital method which takes the id value 
+	Hospital getHospital(Long id); // getHospital method that passes the Long value for the id
+	void deleteHospital(Long id); // DeleteHospital method which takes the id value 
 	void updateHospital(Hospital hospital); // UpdateHospital method which takes in the hospital object
-	int getMaxHospitalID();	// Method that returns the maxhospitalid
+	Long getMaxHospitalID();	// Method that returns the maxhospitalid
 	List<Hospital> getAllHospitals();	// Map object that has a vector of Integer and String values which getsAllHospitals
-	List<ReportsHospital> getAllHospitalsReports(int hospitalId);
+	List<ReportsHospital> getAllHospitalsReports(Long hospitalId);
 	//Map<Integer, String> getHRP();
-	Map<Integer, String> getAllRegions();
+	Map<Long, String> getAllRegions();
 }
